+++ resource/csdk/stack/samples/linux/SimpleClientServer/ocserverbasicops.cpp
@@ -345,7 +345,7 @@ int main(int /*argc*/, char* /*argv*/[])
     return 0;
 }
 
-int createLEDResource (char *uri, LEDResource *ledResource, bool resourceState, int resourcePower)
+int createLEDResource (char *uri, LEDResource *ledResource, bool resourceState, int64_t resourcePower)
 {
     if (!uri)
     {
